Refactor user access token settings into it's own component and add a button spinner
