+++ components/user_settings/security/user_settings_security.jsx
@@ -3,28 +3,23 @@
 
 import PropTypes from 'prop-types';
 import React from 'react';
-import {FormattedDate, FormattedHTMLMessage, FormattedMessage, FormattedTime} from 'react-intl';
+import {FormattedDate, FormattedMessage, FormattedTime} from 'react-intl';
 import {Link} from 'react-router-dom';
-import * as UserUtils from 'mattermost-redux/utils/user_utils';
 
 import {browserHistory} from 'utils/browser_history';
-import {trackEvent} from 'actions/diagnostics_actions.jsx';
 import {deactivateMfa, deauthorizeOAuthApp, getAuthorizedApps, updatePassword} from 'actions/user_actions.jsx';
 import PreferenceStore from 'stores/preference_store.jsx';
 import Constants from 'utils/constants.jsx';
-import {isMobile} from 'utils/user_agent.jsx';
 import * as Utils from 'utils/utils.jsx';
 import icon50 from 'images/icon50x50.png';
 import AccessHistoryModal from 'components/access_history_modal';
 import ActivityLogModal from 'components/activity_log_modal';
-import ConfirmModal from 'components/confirm_modal.jsx';
 import SettingItemMax from 'components/setting_item_max.jsx';
 import SettingItemMin from 'components/setting_item_min.jsx';
 import ToggleModalButton from 'components/toggle_modal_button.jsx';
 
-const TOKEN_CREATING = 'creating';
-const TOKEN_CREATED = 'created';
-const TOKEN_NOT_CREATING = 'not_creating';
+import UserAccessTokenSection from './user_access_token_section';
+
 const SECTION_MFA = 'mfa';
 const SECTION_PASSWORD = 'password';
 const SECTION_SIGNIN = 'signin';
@@ -40,11 +35,6 @@ export default class SecurityTab extends React.Component {
         closeModal: PropTypes.func.isRequired,
         collapseModal: PropTypes.func.isRequired,
         setEnforceFocus: PropTypes.func.isRequired,
-
-        /*
-         * The personal access tokens for the user
-         */
-        userAccessTokens: PropTypes.object,
 
         /*
          * Set if access tokens are enabled and this user can use them
@@ -91,36 +81,6 @@ export default class SecurityTab extends React.Component {
 
         actions: PropTypes.shape({
             getMe: PropTypes.func.isRequired,
-
-            /*
-             * Function to get personal access tokens for a user
-             */
-            getUserAccessTokensForUser: PropTypes.func.isRequired,
-
-            /*
-             * Function to create a personal access token
-             */
-            createUserAccessToken: PropTypes.func.isRequired,
-
-            /*
-             * Function to revoke a personal access token
-             */
-            revokeUserAccessToken: PropTypes.func.isRequired,
-
-            /*
-             * Function to activate a personal access token
-             */
-            enableUserAccessToken: PropTypes.func.isRequired,
-
-            /*
-             * Function to deactivate a personal access token
-             */
-            disableUserAccessToken: PropTypes.func.isRequired,
-
-            /*
-             * Function to clear personal access tokens locally
-             */
-            clearUserAccessTokens: PropTypes.func.isRequired,
         }).isRequired,
     }
 
@@ -138,7 +98,6 @@ export default class SecurityTab extends React.Component {
             passwordError: '',
             serverError: '',
             tokenError: '',
-            showConfirmModal: false,
             authService: this.props.user.auth_service,
             savingPassword: false,
         };
@@ -154,12 +113,6 @@ export default class SecurityTab extends React.Component {
                     this.setState({serverError: err.message}); //eslint-disable-line react/no-did-mount-set-state
                 }
             );
-        }
-
-        if (this.props.canUseAccessTokens) {
-            this.props.actions.clearUserAccessTokens();
-            const userId = this.props.user ? this.props.user.id : '';
-            this.props.actions.getUserAccessTokensForUser(userId, 0, 200);
         }
     }
 
@@ -280,6 +233,7 @@ export default class SecurityTab extends React.Component {
             switch (this.props.activeSection) {
             case SECTION_MFA:
             case SECTION_SIGNIN:
+            case SECTION_TOKENS:
             case SECTION_APPS:
                 this.setState({
                     serverError: null,
@@ -292,14 +246,6 @@ export default class SecurityTab extends React.Component {
                     confirmPassword: '',
                     serverError: null,
                     passwordError: null,
-                });
-                break;
-            case SECTION_TOKENS:
-                this.setState({
-                    newToken: null,
-                    tokenCreationState: TOKEN_NOT_CREATING,
-                    serverError: null,
-                    tokenError: '',
                 });
                 break;
             default:
@@ -1014,436 +960,6 @@ export default class SecurityTab extends React.Component {
         );
     }
 
-    startCreatingToken = () => {
-        this.setState({tokenCreationState: TOKEN_CREATING});
-    }
-
-    stopCreatingToken = () => {
-        this.setState({tokenCreationState: TOKEN_NOT_CREATING});
-    }
-
-    handleCreateToken = async () => {
-        this.handleCancelConfirm();
-
-        const description = this.refs.newtokendescription ? this.refs.newtokendescription.value : '';
-
-        if (description === '') {
-            this.setState({tokenError: Utils.localizeMessage('user.settings.tokens.nameRequired', 'Please enter a description.')});
-            return;
-        }
-
-        this.setState({tokenError: ''});
-
-        const userId = this.props.user ? this.props.user.id : '';
-        const {data, error} = await this.props.actions.createUserAccessToken(userId, description);
-
-        if (data) {
-            this.setState({tokenCreationState: TOKEN_CREATED, newToken: data});
-        } else if (error) {
-            this.setState({serverError: error.message});
-        }
-    }
-
-    handleCancelConfirm = () => {
-        this.setState({
-            showConfirmModal: false,
-            confirmTitle: null,
-            confirmMessage: null,
-            confirmButton: null,
-            confirmComplete: null,
-        });
-    }
-
-    confirmCreateToken = () => {
-        if (UserUtils.isSystemAdmin(this.props.user.roles)) {
-            this.setState({
-                showConfirmModal: true,
-                confirmTitle: (
-                    <FormattedMessage
-                        id='user.settings.tokens.confirmCreateTitle'
-                        defaultMessage='Create System Admin Personal Access Token'
-                    />
-                ),
-                confirmMessage: (
-                    <div className='alert alert-danger'>
-                        <FormattedHTMLMessage
-                            id='user.settings.tokens.confirmCreateMessage'
-                            defaultMessage='You are generating a personal access token with System Admin permissions. Are you sure want to create this token?'
-                        />
-                    </div>
-                ),
-                confirmButton: (
-                    <FormattedMessage
-                        id='user.settings.tokens.confirmCreateButton'
-                        defaultMessage='Yes, Create'
-                    />
-                ),
-                confirmComplete: () => {
-                    this.handleCreateToken();
-                    trackEvent('settings', 'system_admin_create_user_access_token');
-                },
-            });
-
-            return;
-        }
-
-        this.handleCreateToken();
-    }
-
-    saveTokenKeyPress = (e) => {
-        if (Utils.isKeyPressed(e, Constants.KeyCodes.ENTER)) {
-            this.confirmCreateToken();
-        }
-    }
-
-    confirmRevokeToken = (tokenId) => {
-        const token = this.props.userAccessTokens[tokenId];
-
-        this.setState({
-            showConfirmModal: true,
-            confirmTitle: (
-                <FormattedMessage
-                    id='user.settings.tokens.confirmDeleteTitle'
-                    defaultMessage='Delete Token?'
-                />
-            ),
-            confirmMessage: (
-                <div className='alert alert-danger'>
-                    <FormattedHTMLMessage
-                        id='user.settings.tokens.confirmDeleteMessage'
-                        defaultMessage='Any integrations using this token will no longer be able to access the Mattermost API. You cannot undo this action. <br /><br />Are you sure want to delete the {description} token?'
-                        values={{
-                            description: token.description,
-                        }}
-                    />
-                </div>
-            ),
-            confirmButton: (
-                <FormattedMessage
-                    id='user.settings.tokens.confirmDeleteButton'
-                    defaultMessage='Yes, Delete'
-                />
-            ),
-            confirmComplete: () => {
-                this.revokeToken(tokenId);
-                trackEvent('settings', 'revoke_user_access_token');
-            },
-        });
-    }
-
-    revokeToken = async (tokenId) => {
-        const {error} = await this.props.actions.revokeUserAccessToken(tokenId);
-        if (error) {
-            this.setState({serverError: error.message});
-        }
-        this.handleCancelConfirm();
-    }
-
-    activateToken = async (tokenId) => {
-        const {error} = await this.props.actions.enableUserAccessToken(tokenId);
-        if (error) {
-            this.setState({serverError: error.message});
-        } else {
-            trackEvent('settings', 'activate_user_access_token');
-        }
-    }
-
-    deactivateToken = async (tokenId) => {
-        const {error} = await this.props.actions.disableUserAccessToken(tokenId);
-        if (error) {
-            this.setState({serverError: error.message});
-        } else {
-            trackEvent('settings', 'deactivate_user_access_token');
-        }
-    }
-
-    createTokensSection = () => {
-        let tokenListClass = '';
-
-        if (this.props.activeSection === SECTION_TOKENS) {
-            const tokenList = [];
-            Object.values(this.props.userAccessTokens).forEach((token) => {
-                if (this.state.newToken && this.state.newToken.id === token.id) {
-                    return;
-                }
-
-                let activeLink;
-                let activeStatus;
-
-                if (token.is_active) {
-                    activeLink = (
-                        <a
-                            name={token.id + '_deactivate'}
-                            href='#'
-                            onClick={(e) => {
-                                e.preventDefault();
-                                this.deactivateToken(token.id);
-                            }}
-                        >
-                            <FormattedMessage
-                                id='user.settings.tokens.deactivate'
-                                defaultMessage='Deactivate'
-                            />
-                        </a>);
-                } else {
-                    activeStatus = (
-                        <span className='has-error setting-box__inline-error'>
-                            <FormattedMessage
-                                id='user.settings.tokens.deactivatedWarning'
-                                defaultMessage='(Inactive)'
-                            />
-                        </span>
-                    );
-                    activeLink = (
-                        <a
-                            name={token.id + '_activate'}
-                            href='#'
-                            onClick={(e) => {
-                                e.preventDefault();
-                                this.activateToken(token.id);
-                            }}
-                        >
-                            <FormattedMessage
-                                id='user.settings.tokens.activate'
-                                defaultMessage='Activate'
-                            />
-                        </a>
-                    );
-                }
-
-                tokenList.push(
-                    <div
-                        key={token.id}
-                        className='setting-box__item'
-                    >
-                        <div className='whitespace--nowrap overflow--ellipsis'>
-                            <FormattedMessage
-                                id='user.settings.tokens.tokenDesc'
-                                defaultMessage='Token Description: '
-                            />
-                            {token.description}
-                            {activeStatus}
-                        </div>
-                        <div className='setting-box__token-id whitespace--nowrap overflow--ellipsis'>
-                            <FormattedMessage
-                                id='user.settings.tokens.tokenId'
-                                defaultMessage='Token ID: '
-                            />
-                            {token.id}
-                        </div>
-                        <div>
-                            {activeLink}
-                            {' - '}
-                            <a
-                                name={token.id + '_delete'}
-                                href='#'
-                                onClick={(e) => {
-                                    e.preventDefault();
-                                    this.confirmRevokeToken(token.id);
-                                }}
-                            >
-                                <FormattedMessage
-                                    id='user.settings.tokens.delete'
-                                    defaultMessage='Delete'
-                                />
-                            </a>
-                        </div>
-                        <hr className='margin-bottom margin-top x2'/>
-                    </div>
-                );
-            });
-
-            let noTokenText;
-            if (tokenList.length === 0) {
-                noTokenText = (
-                    <FormattedMessage
-                        key='notokens'
-                        id='user.settings.tokens.userAccessTokensNone'
-                        defaultMessage='No personal access tokens.'
-                    />
-                );
-            }
-
-            let extraInfo;
-            if (isMobile()) {
-                extraInfo = (
-                    <span>
-                        <FormattedHTMLMessage
-                            id='user.settings.tokens.description_mobile'
-                            defaultMessage='<a href="https://about.mattermost.com/default-user-access-tokens" target="_blank">Personal access tokens</a> function similarly to session tokens and can be used by integrations to <a href="https://about.mattermost.com/default-api-authentication" target="_blank">authenticate against the REST API</a>. Create new tokens on your desktop.'
-                        />
-                    </span>
-                );
-            } else {
-                extraInfo = (
-                    <span>
-                        <FormattedHTMLMessage
-                            id='user.settings.tokens.description'
-                            defaultMessage='<a href="https://about.mattermost.com/default-user-access-tokens" target="_blank">Personal access tokens</a> function similarly to session tokens and can be used by integrations to <a href="https://about.mattermost.com/default-api-authentication" target="_blank">authenticate against the REST API</a>.'
-                        />
-                    </span>
-                );
-            }
-
-            let newTokenSection;
-            if (this.state.tokenCreationState === TOKEN_CREATING) {
-                newTokenSection = (
-                    <div className='padding-left x2'>
-                        <div className='row'>
-                            <label className='col-sm-auto control-label padding-right x2'>
-                                <FormattedMessage
-                                    id='user.settings.tokens.name'
-                                    defaultMessage='Token Description: '
-                                />
-                            </label>
-                            <div className='col-sm-5'>
-                                <input
-                                    autoFocus={true}
-                                    ref='newtokendescription'
-                                    className='form-control'
-                                    type='text'
-                                    maxLength={64}
-                                    onKeyPress={this.saveTokenKeyPress}
-                                />
-                            </div>
-                        </div>
-                        <div>
-                            <div className='padding-top x2'>
-                                <FormattedMessage
-                                    id='user.settings.tokens.nameHelp'
-                                    defaultMessage='Enter a description for your token to remember what it does.'
-                                />
-                            </div>
-                            <div>
-                                <label
-                                    id='clientError'
-                                    className='has-error margin-top margin-bottom'
-                                >
-                                    {this.state.tokenError}
-                                </label>
-                            </div>
-                            <button
-                                className='btn btn-primary'
-                                onClick={this.confirmCreateToken}
-                            >
-                                <FormattedMessage
-                                    id='user.settings.tokens.save'
-                                    defaultMessage='Save'
-                                />
-                            </button>
-                            <button
-                                className='btn btn-default'
-                                onClick={this.stopCreatingToken}
-                            >
-                                <FormattedMessage
-                                    id='user.settings.tokens.cancel'
-                                    defaultMessage='Cancel'
-                                />
-                            </button>
-                        </div>
-                    </div>
-                );
-            } else if (this.state.tokenCreationState === TOKEN_CREATED) {
-                if (tokenList.length === 0) {
-                    tokenListClass = ' hidden';
-                }
-
-                newTokenSection = (
-                    <div
-                        className='alert alert-warning'
-                    >
-                        <i className='fa fa-warning margin-right'/>
-                        <FormattedMessage
-                            id='user.settings.tokens.copy'
-                            defaultMessage="Please copy the access token below. You won't be able to see it again!"
-                        />
-                        <br/>
-                        <br/>
-                        <div className='whitespace--nowrap overflow--ellipsis'>
-                            <FormattedMessage
-                                id='user.settings.tokens.name'
-                                defaultMessage='Token Description: '
-                            />
-                            {this.state.newToken.description}
-                        </div>
-                        <div className='whitespace--nowrap overflow--ellipsis'>
-                            <FormattedMessage
-                                id='user.settings.tokens.id'
-                                defaultMessage='Token ID: '
-                            />
-                            {this.state.newToken.id}
-                        </div>
-                        <strong className='word-break--all'>
-                            <FormattedMessage
-                                id='user.settings.tokens.token'
-                                defaultMessage='Access Token: '
-                            />
-                            {this.state.newToken.token}
-                        </strong>
-                    </div>
-                );
-            } else {
-                newTokenSection = (
-                    <a
-                        className='btn btn-primary'
-                        href='#'
-                        onClick={this.startCreatingToken}
-                    >
-                        <FormattedMessage
-                            id='user.settings.tokens.create'
-                            defaultMessage='Create New Token'
-                        />
-                    </a>
-                );
-            }
-
-            const inputs = [];
-            inputs.push(
-                <div
-                    key='tokensSetting'
-                    className='padding-top'
-                >
-                    <div key='tokenList'>
-                        <div className={'alert alert-transparent' + tokenListClass}>
-                            {tokenList}
-                            {noTokenText}
-                        </div>
-                        {newTokenSection}
-                    </div>
-                </div>
-            );
-
-            return (
-                <SettingItemMax
-                    title={Utils.localizeMessage('user.settings.tokens.title', 'Personal Access Tokens')}
-                    inputs={inputs}
-                    extraInfo={extraInfo}
-                    infoPosition='top'
-                    serverError={this.state.serverError}
-                    updateSection={this.handleUpdateSection}
-                    width='full'
-                    cancelButtonText={
-                        <FormattedMessage
-                            id='user.settings.security.close'
-                            defaultMessage='Close'
-                        />
-                    }
-                />
-            );
-        }
-
-        const describe = Utils.localizeMessage('user.settings.tokens.clickToEdit', "Click 'Edit' to manage your personal access tokens");
-
-        return (
-            <SettingItemMin
-                title={Utils.localizeMessage('user.settings.tokens.title', 'Personal Access Tokens')}
-                describe={describe}
-                section={SECTION_TOKENS}
-                updateSection={this.handleUpdateSection}
-            />
-        );
-    }
-
     render() {
         const user = this.props.user;
 
@@ -1477,7 +993,13 @@ export default class SecurityTab extends React.Component {
 
         let tokensSection;
         if (this.props.canUseAccessTokens) {
-            tokensSection = this.createTokensSection();
+            tokensSection = (
+                <UserAccessTokenSection
+                    user={this.props.user}
+                    active={this.props.activeSection === SECTION_TOKENS}
+                    updateSection={this.handleUpdateSection}
+                />
+            );
         }
 
         return (
@@ -1548,14 +1070,6 @@ export default class SecurityTab extends React.Component {
                         />
                     </ToggleModalButton>
                 </div>
-                <ConfirmModal
-                    title={this.state.confirmTitle}
-                    message={this.state.confirmMessage}
-                    confirmButtonText={this.state.confirmButton}
-                    show={this.state.showConfirmModal}
-                    onConfirm={this.state.confirmComplete || (() => {})} //eslint-disable-line no-empty-function
-                    onCancel={this.handleCancelConfirm}
-                />
             </div>
         );
     }
